refactor: 상품 검색 기능 개선을 위한 useProductSearch 훅 추가 및 App 컴포넌트 수정
